backendSuchus/requirements.txt
Django==5.2.4
djangorestframework==3.14.0
django-cors-headers==4.3.1djangorestframework-simplejwt==5.3.1python-dotenv==1.0.0
boto3==1.26.137
cloudinary==1.33.0
django-cloudinary-storage==0.3.0
